Plugin sqlite_visu2_8: Test fix for _series issue with tend != now
If _series is called with tend != now, the returned series contains items that are not ordered.
This causes issues eg with SmartVísu Plots.

This commit adds an addtional method _series2 which has a slightly different handling that
should avoid the issue. The method is added as method to items as <item>.series2. Method
interface is the same than <item>.series.

When comparing results please bear in mind that the timestamp of the first element is always <tstart>,
the timestamp of the last element is always <tend>